pyproject.toml
[tool.poetry]
name = "smartfoosball"
version = "0.1.0"
description = "Very smart table football table"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.9"
opencv-python = "^4.10.0.84"
pyserial = "^3.5"
waitress = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
